refactor(lib): deepen Overpass snapshot pipeline
Split lib/osm.ts into three focused modules for better separation of concerns:

- lib/overpass/transport.ts (private): HTTP transport layer with executeOverpassQuery,
  convertOverpassToGeoJSON, getUserAgent, and test guard
- lib/dataset-snapshot.ts (public deep): fetchDatasetSnapshot with narrow public interface
  (DatasetSnapshot, DatasetStats types)
- lib/area-boundary.ts (enriched): absorbs fetchOsmRelationData, now owns all OSM
  relation-level knowledge

Benefits:
- Locality: all Overpass HTTP logic concentrated in one private module
- Leverage: callers see one interface (fetchDatasetSnapshot) instead of 5 low-level functions
- Seams: enables in-memory test adapter for dataset snapshots

Deleted lib/osm.ts. Updated 9 callers to import from new modules.
Migrated osm.test.ts → dataset-snapshot.test.ts. Updated nominatim.ts to import
getUserAgent from transport.

All tests pass (6/6 dataset-snapshot tests).
