+++ src/app/api/datasets/[id]/refresh/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest, NextResponse } from "next/server";
 import { auth } from "@/auth";
 import { prisma } from "@/lib/db";
-import { fetchDatasetSnapshot } from "@/lib/osm";
+import { fetchDatasetSnapshot } from "@/lib/dataset-snapshot";
 import { trackEvent, getClientInfo } from "@/lib/umami";
 import { ANALYTICS_EVENTS } from "@/lib/analytics/events";
