+++ src/lib/dataset-operations.ts
@@ -2,7 +2,8 @@ import { prisma } from "@/lib/db";
 import { getAreaDetailsById } from "@/lib/nominatim";
 import { resolveTemplate } from "@/lib/template-resolver";
 import { resolveTemplateForLocale } from "@/lib/template-locale";
-import { fetchOsmRelationData, fetchDatasetSnapshot } from "@/lib/osm";
+import { fetchOsmRelationData } from "@/lib/area-boundary";
+import { fetchDatasetSnapshot } from "@/lib/dataset-snapshot";
 import { Prisma } from "@prisma/client";
 import { trackEvent } from "@/lib/umami";
 import { ANALYTICS_EVENTS } from "@/lib/analytics/events";
